requirements.txt
numpy
matplotlib
Pillow
torch
torchvision
scikit-learn
tqdm
pandas
ml_metrics
jupyter
